refactor: Переписывает стили и удаляет моковые данные для предложений
