--- src/mocks/offers-mock.ts
@@ -1,1318 +1,0 @@
-import {OfferType} from '../shared/types';
-
-const OFFERS_LIST_MOCK: OfferType[] = [
-  {
-    'id': 'a0a5c15f-50be-4a09-bb62-0f8af62f5625',
-    'title': 'Tile House',
-    'type': 'room',
-    'price': 131,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/1.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.868610000000004,
-      'longitude': 2.342499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 2.3
-  },
-  {
-    'id': 'df4bdd1d-9030-4a67-a646-adf37237a36a',
-    'title': 'Amazing and Extremely Central Flat',
-    'type': 'room',
-    'price': 140,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/17.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.858610000000006,
-      'longitude': 2.330499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.7
-  },
-  {
-    'id': 'dc5a063e-40bf-4348-b7b5-86b9662dc83e',
-    'title': 'Tile House',
-    'type': 'room',
-    'price': 199,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/15.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.834610000000005,
-      'longitude': 2.335499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.5
-  },
-  {
-    'id': 'ee44c6d3-fd98-4294-8f6e-3262802dfb40',
-    'title': 'Perfectly located Castro',
-    'type': 'hotel',
-    'price': 392,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/7.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.85761,
-      'longitude': 2.358499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.7
-  },
-  {
-    'id': 'b077788a-a4b8-4567-8f43-63376a03fbcb',
-    'title': 'Loft Studio in the Central Area',
-    'type': 'house',
-    'price': 813,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/18.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.87561,
-      'longitude': 2.375499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.3
-  },
-  {
-    'id': 'acddc3f9-0007-4c69-ba6f-50b46ea42333',
-    'title': 'Loft Studio in the Central Area',
-    'type': 'hotel',
-    'price': 322,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.87961000000001,
-      'longitude': 2.353499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.4
-  },
-  {
-    'id': '95497202-e973-425d-9033-0a42162112d2',
-    'title': 'Loft Studio in the Central Area',
-    'type': 'room',
-    'price': 100,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.834610000000005,
-      'longitude': 2.364499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.3
-  },
-  {
-    'id': 'c112f49d-e120-4ad1-98f4-86b27d276b89',
-    'title': 'The house among olive ',
-    'type': 'apartment',
-    'price': 427,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.837610000000005,
-      'longitude': 2.3454990000000002,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 1
-  },
-  {
-    'id': '6dd49eb2-769d-4af5-b50b-327477b178a6',
-    'title': 'House in countryside',
-    'type': 'hotel',
-    'price': 202,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/12.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.84761,
-      'longitude': 2.356499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.2
-  },
-  {
-    'id': '2c37e5de-81dd-4110-bb1e-5345eefb6ec8',
-    'title': 'The Joshua Tree House',
-    'type': 'room',
-    'price': 190,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/18.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.862610000000004,
-      'longitude': 2.369499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.8
-  },
-  {
-    'id': '8e782827-a2c8-4493-9124-c5b83ab6ea4b',
-    'title': 'Beautiful & luxurious apartment at great location',
-    'type': 'hotel',
-    'price': 351,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/19.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.83861,
-      'longitude': 2.350499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.7
-  },
-  {
-    'id': 'f83e2002-d076-4ee4-9b7e-da723722c2ec',
-    'title': 'Canal View Prinsengracht',
-    'type': 'room',
-    'price': 139,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/6.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.861610000000006,
-      'longitude': 2.340499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 2
-  },
-  {
-    'id': '43ad31ab-2659-4148-a034-4d0497217793',
-    'title': 'Perfectly located Castro',
-    'type': 'house',
-    'price': 319,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/6.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.87861,
-      'longitude': 2.357499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.4
-  },
-  {
-    'id': 'c37af369-b833-4d27-b04c-a05d6a045bb6',
-    'title': 'The house among olive ',
-    'type': 'room',
-    'price': 222,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.877610000000004,
-      'longitude': 2.333499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.6
-  },
-  {
-    'id': 'b8f91f8b-863c-4f2c-b7ae-03126b4af95b',
-    'title': 'Perfectly located Castro',
-    'type': 'house',
-    'price': 247,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/17.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.83961,
-      'longitude': 2.342499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.9
-  },
-  {
-    'id': '7b8431f1-52f3-4551-84a3-0dd6996795fc',
-    'title': 'Loft Studio in the Central Area',
-    'type': 'hotel',
-    'price': 287,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/1.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.865610000000004,
-      'longitude': 2.350499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3
-  },
-  {
-    'id': '8fb85a86-b6f1-46d6-847b-9ecafc8ff433',
-    'title': 'The Joshua Tree House',
-    'type': 'hotel',
-    'price': 160,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.846610000000005,
-      'longitude': 2.374499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 4.4
-  },
-  {
-    'id': 'f07b25c6-dba1-4712-83be-b7423673b044',
-    'title': 'The Joshua Tree House',
-    'type': 'house',
-    'price': 309,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/20.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.837610000000005,
-      'longitude': 2.364499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.1
-  },
-  {
-    'id': '8d9a2672-5f47-47b5-a54b-be600233e5f3',
-    'title': 'Nice, cozy, warm big bed apartment',
-    'type': 'apartment',
-    'price': 493,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/13.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.843610000000005,
-      'longitude': 2.338499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 1
-  },
-  {
-    'id': '4e68cf9a-f722-4b24-b8a8-a7df9bfef8ef',
-    'title': 'Nice, cozy, warm big bed apartment',
-    'type': 'hotel',
-    'price': 490,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/1.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.84461,
-      'longitude': 2.374499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.3
-  },
-
-  {
-    'id': 'd689221d-9478-47d2-8197-f8773a4d5a75',
-    'title': 'The Pondhouse - A Magical Place',
-    'type': 'apartment',
-    'price': 314,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/13.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.918461,
-      'longitude': 6.969974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.2
-  },
-  {
-    'id': '7da077d7-7d41-4038-abdb-4912e8644aae',
-    'title': 'Loft Studio in the Central Area',
-    'type': 'apartment',
-    'price': 179,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/20.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.957361,
-      'longitude': 6.9509739999999995,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.9
-  },
-  {
-    'id': '06bb4a3c-0e3f-4777-9348-58fec4047538',
-    'title': 'Perfectly located Castro',
-    'type': 'room',
-    'price': 148,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/17.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.951361,
-      'longitude': 6.944974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 4.9
-  },
-  {
-    'id': 'f2016e29-73ef-45c5-be1c-8000ce77b32d',
-    'title': 'Tile House',
-    'type': 'room',
-    'price': 250,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/16.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.959361,
-      'longitude': 6.978974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 1.4
-  },
-  {
-    'id': 'fc4cdc02-f77f-41cf-bcb0-8e31ee9b53ad',
-    'title': 'Penthouse, 4-5 rooms + 5 balconies',
-    'type': 'hotel',
-    'price': 130,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.932361,
-      'longitude': 6.960974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 4.3
-  },
-  {
-    'id': '71ec1dcb-b38d-4237-be84-85e7124a8134',
-    'title': 'Canal View Prinsengracht',
-    'type': 'hotel',
-    'price': 109,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/13.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.954361,
-      'longitude': 6.982974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.7
-  },
-  {
-    'id': '60ba614f-83cd-4af5-9d12-95eb7ac97804',
-    'title': 'Perfectly located Castro',
-    'type': 'hotel',
-    'price': 382,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/11.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.945361,
-      'longitude': 6.962974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 1.1
-  },
-  {
-    'id': 'ddc4dceb-a85b-4e75-909b-54750c24fb5b',
-    'title': 'Loft Studio in the Central Area',
-    'type': 'house',
-    'price': 193,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/9.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.917361,
-      'longitude': 6.977974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.6
-  },
-  {
-    'id': '0baa919c-facb-4336-9db0-03d8f8753c93',
-    'title': 'The house among olive ',
-    'type': 'house',
-    'price': 893,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/9.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.945361,
-      'longitude': 6.935974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.8
-  },
-  {
-    'id': '3f9aa4a8-ef03-413d-aca7-8e2d6fe73f0a',
-    'title': 'Waterfront with extraordinary view',
-    'type': 'house',
-    'price': 475,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/8.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.36554,
-      'longitude': 4.911976,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.8
-  },
-  {
-    'id': 'afb549ff-1ae3-485f-baa5-a3a035a69a7b',
-    'title': 'Perfectly located Castro',
-    'type': 'house',
-    'price': 797,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/8.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.35054,
-      'longitude': 4.908976,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.3
-  },
-  {
-    'id': 'f2f86c4d-109d-445b-8395-a93af5fbe875',
-    'title': 'Amazing and Extremely Central Flat',
-    'type': 'apartment',
-    'price': 309,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/11.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.37154,
-      'longitude': 4.889976,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.6
-  },
-  {
-    'id': '1a53399c-4eff-4221-bb53-2fbb0de4539f',
-    'title': 'Beautiful & luxurious apartment at great location',
-    'type': 'hotel',
-    'price': 302,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.364540000000005,
-      'longitude': 4.9019759999999994,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3
-  },
-  {
-    'id': '46c27c42-1e9d-4ae9-9dd7-23730f3469d8',
-    'title': 'House in countryside',
-    'type': 'hotel',
-    'price': 111,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/17.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.36354,
-      'longitude': 4.911976,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.5
-  },
-  {
-    'id': 'ab34e9a7-4a58-4e95-852b-a60500b313c4',
-    'title': 'Amazing and Extremely Central Flat',
-    'type': 'apartment',
-    'price': 423,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.36354,
-      'longitude': 4.889976,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.1
-  },
-  {
-    'id': 'e7372574-9441-45e3-bd2d-44c9ad2abafd',
-    'title': 'The Pondhouse - A Magical Place',
-    'type': 'room',
-    'price': 250,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/4.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.388540000000006,
-      'longitude': 4.899976,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.4
-  },
-  {
-    'id': 'f22c8821-b6b7-437e-b958-6cffb83cc147',
-    'title': 'Amazing and Extremely Central Flat',
-    'type': 'apartment',
-    'price': 231,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/20.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.36954,
-      'longitude': 4.914976,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.6
-  },
-  {
-    'id': '47ec9d57-38b7-4096-b601-3665ba8a5287',
-    'title': 'Wood and stone place',
-    'type': 'house',
-    'price': 949,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/13.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.37554,
-      'longitude': 4.9019759999999994,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.2
-  },
-  {
-    'id': '02b1c2e2-69a4-4b64-b01a-bc3574a7b074',
-    'title': 'The Joshua Tree House',
-    'type': 'apartment',
-    'price': 255,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/15.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.385540000000006,
-      'longitude': 4.886976,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.7
-  },
-  {
-    'id': 'ac744960-230e-4c16-b3c0-2bf3fd38e62d',
-    'title': 'Tile House',
-    'type': 'house',
-    'price': 349,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/17.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.565341000000004,
-      'longitude': 9.978654,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.3
-  },
-  {
-    'id': 'a710e1f4-f452-4ab5-a92c-84d950993870',
-    'title': 'The house among olive ',
-    'type': 'apartment',
-    'price': 236,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/15.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.570341000000006,
-      'longitude': 9.975654,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.7
-  },
-  {
-    'id': 'd57ccbff-48bc-449e-94f6-c78badb8c572',
-    'title': 'Canal View Prinsengracht',
-    'type': 'house',
-    'price': 576,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/17.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.558341000000006,
-      'longitude': 10.001654,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.7
-  },
-  {
-    'id': '8d6e04e9-c00e-48cb-813d-4203889d7f67',
-    'title': 'Nice, cozy, warm big bed apartment',
-    'type': 'hotel',
-    'price': 422,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/8.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.573341000000006,
-      'longitude': 10.009654000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.2
-  },
-  {
-    'id': '7abda6d5-6a50-4b04-858a-863a489edb6d',
-    'title': 'Tile House',
-    'type': 'apartment',
-    'price': 326,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/1.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.573341000000006,
-      'longitude': 9.994654,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.2
-  },
-  {
-    'id': '2f9c5423-4cdf-47a5-afbc-43b52a0a0026',
-    'title': 'The house among olive ',
-    'type': 'hotel',
-    'price': 204,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/17.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.528341000000005,
-      'longitude': 9.982654,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.9
-  },
-  {
-    'id': '30d059f8-c256-4766-be4c-7862cab5f491',
-    'title': 'Perfectly located Castro',
-    'type': 'hotel',
-    'price': 387,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.540341000000005,
-      'longitude': 10.025654000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.4
-  },
-  {
-    'id': '73238a05-3102-41fe-a40f-10f01966d8d5',
-    'title': 'Amazing and Extremely Central Flat',
-    'type': 'room',
-    'price': 191,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.555341000000006,
-      'longitude': 9.975654,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.5
-  },
-  {
-    'id': 'dc31a79c-35f5-4f65-89ac-7ae4b36009db',
-    'title': 'The Joshua Tree House',
-    'type': 'room',
-    'price': 214,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/16.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.534341000000005,
-      'longitude': 9.998654,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 4.8
-  },
-  {
-    'id': '29e35660-7d91-40f6-b1dc-618e8ea6318f',
-    'title': 'Waterfront with extraordinary view',
-    'type': 'room',
-    'price': 205,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.563341,
-      'longitude': 9.975654,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.3
-  },
-  {
-    'id': '2ec9ebe4-a878-4147-9bda-d55a93d13c17',
-    'title': 'Loft Studio in the Central Area',
-    'type': 'hotel',
-    'price': 153,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/6.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.574341000000004,
-      'longitude': 10.022654000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.1
-  },
-  {
-    'id': '16d95b59-2305-46f2-ab69-f085f79ace6f',
-    'title': 'Canal View Prinsengracht',
-    'type': 'house',
-    'price': 807,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/6.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.573341000000006,
-      'longitude': 10.025654000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.4
-  },
-  {
-    'id': '52d44aff-fa09-4827-8715-8f9a4f3d4fb0',
-    'title': 'Tile House',
-    'type': 'house',
-    'price': 815,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/2.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.550341,
-      'longitude': 9.980654000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 4
-  },
-  {
-    'id': '912270f0-1fd1-4f43-b729-131c7f228afa',
-    'title': 'The Joshua Tree House',
-    'type': 'apartment',
-    'price': 242,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/8.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.558341000000006,
-      'longitude': 9.999654000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 2.2
-  },
-  {
-    'id': '2ea32e64-e27c-4677-9839-80d966b49e78',
-    'title': 'Loft Studio in the Central Area',
-    'type': 'room',
-    'price': 173,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/6.jpg',
-    'city': {
-      'name': 'Dusseldorf',
-      'location': {
-        'latitude': 51.225402,
-        'longitude': 6.776314,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 51.239402000000005,
-      'longitude': 6.756314000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.8
-  },
-  {
-    'id': 'e167c342-296a-4139-b86e-bb26465c3d43',
-    'title': 'Perfectly located Castro',
-    'type': 'hotel',
-    'price': 123,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/20.jpg',
-    'city': {
-      'name': 'Dusseldorf',
-      'location': {
-        'latitude': 51.225402,
-        'longitude': 6.776314,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 51.243402,
-      'longitude': 6.791314,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 2.7
-  },
-  {
-    'id': '68734aae-04e4-457f-81cf-b43048d1fc61',
-    'title': 'Nice, cozy, warm big bed apartment',
-    'type': 'house',
-    'price': 926,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/12.jpg',
-    'city': {
-      'name': 'Dusseldorf',
-      'location': {
-        'latitude': 51.225402,
-        'longitude': 6.776314,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 51.214402,
-      'longitude': 6.764314000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 1.6
-  },
-  {
-    'id': '1eea1a90-cf77-4ba6-93aa-fb8fe5692dd0',
-    'title': 'Beautiful & luxurious apartment at great location',
-    'type': 'hotel',
-    'price': 468,
-    'previewImage': 'https://16.design.htmlacademy.pro/static/hotel/13.jpg',
-    'city': {
-      'name': 'Dusseldorf',
-      'location': {
-        'latitude': 51.225402,
-        'longitude': 6.776314,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 51.225402,
-      'longitude': 6.784314,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 1.1
-  }
-];
-
-export default OFFERS_LIST_MOCK;
