+++ src/widgets/header/ui/header.css
@@ -1,20 +1,3 @@
-.header__user-name {
-  position: relative;
-}
-
-.tooltip {
-  display: none;
-  width: 320px;
-  color: #383838;
-  text-align: center;
-  border-radius: 5px;
-  padding: 5px;
-  position: absolute;
-  z-index: 1;
-  bottom: -100%;
-  margin-left: -60px;
-}
-
 .header__nav-link:hover {
   .tooltip {
     display: block;
@@ -41,4 +24,8 @@
 
 .header__nav-link {
   cursor: pointer;
+}
+
+.user__link {
+ display: inherit;
 }
